requirements.txt
fastapi
uvicorn[standard]
pydantic
pyjwt
redis
pywebpush
pika==1.3.2
asgiref
pydantic-settings